+++ src/pages/task/newTask/AddNewTask.tsx
@@ -858,7 +858,7 @@ const AddNewTask = () => {
                       defaultChecked={x.isChecked}
                     />{' '}
                     <label style={{ padding: '0 8px 0 0' }}>{x.name}</label>
-                    {index === 4 && (
+                    {x.name === 'อื่นๆ' && (
                       <>
                         <Input
                           status={validateComma.status}